+++ test/unit/mysql_db_dump/dump_run.py
@@ -29,10 +29,46 @@ import mock
 # Local
 sys.path.append(os.getcwd())
 import mysql_db_dump
-import lib.gen_libs as gen_libs
 import version
 
 __version__ = version.__version__
+
+
+class SubProcess(object):
+
+    """Class:  SubProcess
+
+    Description:  Class which is a representation of the subprocess class.
+
+    Methods:
+        __init__ -> Initialize configuration environment.
+        wait -> subprocess.wait method.
+
+    """
+
+    def __init__(self):
+
+        """Method:  __init__
+
+        Description:  Initialization instance of the ZipFile class.
+
+        Arguments:
+
+        """
+
+        pass
+
+    def wait(self):
+
+        """Method:  wait
+
+        Description:  Mock representation of subprocess.wait method.
+
+        Arguments:
+
+        """
+
+        pass
 
 
 class UnitTest(unittest.TestCase):
@@ -43,9 +79,11 @@ class UnitTest(unittest.TestCase):
 
     Methods:
         setUp -> Initialize testing environment.
+        test_error_file -> Test with error file passed.
         test_compress_true -> Test with compression set to True.
         test_compress_false -> Test with compression set to False.
         test_dump_run -> Test with only default arguments passed.
+        tearDown -> Clean up of unit testing.
 
     """
 
@@ -60,13 +98,30 @@ class UnitTest(unittest.TestCase):
         """
 
         self.dump_cmd = []
-        self.dmp_file = "/dir/path/dump_file.dmp"
+        self.dmp_file = "./test/unit/mysql_db_dump/tmp/test_run_prog"
+        self.err_file = "Error File"
+        self.subp = SubProcess()
+
+    @mock.patch("mysql_db_dump.subprocess.Popen")
+    def test_error_file(self, mock_subp):
+
+        """Function:  test_error_file
+
+        Description:  Test with error file passed.
+
+        Arguments:
+
+        """
+
+        mock_subp.return_value = self.subp
+
+        self.assertFalse(mysql_db_dump.dump_run(self.dump_cmd, self.dmp_file,
+                                                False, errfile=self.err_file))
 
     @mock.patch("mysql_db_dump.gen_libs.compress",
                 mock.Mock(return_value=True))
-    @mock.patch("mysql_db_dump.cmds_gen.run_prog",
-                mock.Mock(return_value=True))
-    def test_compress_true(self):
+    @mock.patch("mysql_db_dump.subprocess.Popen")
+    def test_compress_true(self, mock_subp):
 
         """Function:  test_compress_true
 
@@ -76,12 +131,13 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_subp.return_value = self.subp
+
         self.assertFalse(mysql_db_dump.dump_run(self.dump_cmd, self.dmp_file,
                                                 True))
 
-    @mock.patch("mysql_db_dump.cmds_gen.run_prog",
-                mock.Mock(return_value=True))
-    def test_compress_false(self):
+    @mock.patch("mysql_db_dump.subprocess.Popen")
+    def test_compress_false(self, mock_subp):
 
         """Function:  test_compress_false
 
@@ -91,12 +147,13 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_subp.return_value = self.subp
+
         self.assertFalse(mysql_db_dump.dump_run(self.dump_cmd, self.dmp_file,
                                                 False))
 
-    @mock.patch("mysql_db_dump.cmds_gen.run_prog",
-                mock.Mock(return_value=True))
-    def test_dump_run(self):
+    @mock.patch("mysql_db_dump.subprocess.Popen")
+    def test_dump_run(self, mock_subp):
 
         """Function:  test_dump_run
 
@@ -106,8 +163,23 @@ class UnitTest(unittest.TestCase):
 
         """
 
+        mock_subp.return_value = self.subp
+
         self.assertFalse(mysql_db_dump.dump_run(self.dump_cmd, self.dmp_file,
                                                 False))
+
+    def tearDown(self):
+
+        """Function:  tearDown
+
+        Description:  Clean up of unit testing.
+
+        Arguments:
+
+        """
+
+        if os.path.isfile(self.dmp_file):
+            os.remove(self.dmp_file)
 
 
 if __name__ == "__main__":